msm: ipa: Fix incorrect wakelock handling
For every __pm_stay_awake call, there must be a
__pm_relax call to ensure there is no stale
wakelock is held. with the current logic
__pm_stay_awake is called whenever wakelock ref
cnt is non-zero and released only when it is 0.
This leads to wakelock being held even when
it is not required. Make a change to acquire
wakelock only when it is not held previously.

Change-Id: I8c61c184506c705fd5866de8a606f642c5de3d5b
Acked-by: Chaitanya Pratapa <cpratapa@qti.qualcomm.com>
Signed-off-by: Utkarsh Saxena <usaxena@codeaurora.org>
Signed-off-by: Tiktodz <ewprjkt@proton.me>