usb: gadget: function: f_fs: Add FUNCTIONFS_ENDPOINT_ALLOC ioctl
FUNCTIONFS_ENDPOINT_ALLOC will allocate a buffer of a given
size for an endpoint.

Whenever a synchronous i/o is called that can fit in that buffer,
it is used instead of allocating a new buffer on each call.

The buffer is freed when the epfile is released, a new buffer is
allocated, or the ioctl is called with argument 0.

Change-Id: I54af3205575f2bb83c732bb70907f60e0b785e11
Signed-off-by: Jerry Zhang <zhangjerry@google.com>