thermal: qpnp-adc-tm: Fix aborting suspend
When a low hysteresis for thresholds are set there is a
possibility that low and high thresholds are crossed at
around the same time before the client is notified. Under
such condition there could be a low and high interrupt
received before processing the request and notifying the
client. Check the return value for queued item for
notifying the client and reduce the count used to prevent
suspend if the item is already available in the global queue.

Change-Id: I4bddf6742a0686df57b6c4ce0d964579de01b027
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>
Signed-off-by: Raphiel Rollerscaperers <raphielscape@outlook.com>
idkwhoiam322: Modified for our thermal setup.
Signed-off-by: idkwhoiam322 <snkm1999@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: RyuujiX <saputradenny712@gmail.com>
Signed-off-by: dotkit <dotkit@electrowizard.me>