sched/tune: Rename dynamic_boost parameter to sched_boost
This was confusing to deal with given that it had the same name as the
Dynamic Schedtune Boost framework. It will be more apt to call it
sched_boost given that it was created to work with the sched_boost
feature in CAF devices.

The new tunable can be found in /dev/stune/*/schedtune.sched_boost

Change-Id: I1e6b7225c8820cacd7717ee0ab8060a0c186ce38
Signed-off-by: joshuous <joshuous@gmail.com>
Signed-off-by: dotkit <ewprjkt@proton.me>