ion: Speed up handle idr and rbtree lookup with RCU
Convert the handle idr and rbtree to be RCU-safe in order to speed up
their lookups so that they never block. The idr API is already RCU safe,
so converting it to use RCU is easy; however, the rbtree API isn't, so
change it out for the RCU-safe variant known as rbtree_latch.

Change-Id: Ic4e2d509498df60a5a1962fc8cdf23d2d5ef4e31
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>