msm: kgsl: Wake GPU upon receiving an ioctl rather than upon touch input
Waking the GPU upon touch wastes power when the screen is being touched
in a way that does not induce animation or any actual need for GPU usage.
Instead of preemptively waking the GPU on touch input, wake it up upon
receiving a IOCTL_KGSL_GPU_COMMAND ioctl since it is a sign that the GPU
will soon be needed.

Change-Id: I23f9bde98fcbdb39d68f60b5721a9e880dd92fd6
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: texascake <tiktodz@soulvibe-team.cloud>