socket: ignore v6 endpoints when ipv6 is disabled
The previous commit fixed a memory leak on the send path in the event
that IPv6 is disabled at compile time, but how did a packet even arrive
there to begin with? It turns out we have previously allowed IPv6
endpoints even when IPv6 support is disabled at compile time. This is
awkward and inconsistent. Instead, let's just ignore all things IPv6,
the same way we do other malformed endpoints, in the case where IPv6 is
disabled.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I7def019000309c6d1f9aef695f1d4f63460e59dd
Signed-off-by: texascake <tiktodz@soulvibe-team.cloud>